daemon: move hubble related methods to hubble.go
Let's keep everything hubble related in the same file. This commit does
not bring any functional changes.
